backend/requirements.txt
# CyberXLTR Admin Backend Requirements
# PostgreSQL-based admin system with cross-service sync

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database & ORM (PostgreSQL via async SQLAlchemy)
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Validation & Settings
pydantic>=2.7.4
pydantic-settings>=2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# HTTP Clients & Async (for inter-service communication)
httpx>=0.25.0
aiohttp==3.9.1

# Data Processing
pandas==2.2.0
python-dateutil==2.8.2

# Logging & Monitoring
structlog==24.1.0

# Utilities
python-dotenv==1.0.1
pytz==2024.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
